requirements.txt
python-rapidjson 
matplotlib 
watchdog 
forbiddenfruit 
numpy
scipy